// src/generated/api/catalog/models/productVariantDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * DCTD-UTC Commerce API
 * Contract for storefront and admin applications
 * OpenAPI spec version: 1.0.0
 */
import type { ProductVariantStatus } from './productVariantStatus';
import type { ProductVariantDtoBundle } from './productVariantDtoBundle';

export interface ProductVariantDto {
  /** @pattern ^[1-9][0-9]*$ */
  id: string;
  sku: string;
  barcode?: string;
  name: string;
  /** @minimum 0 */
  weightGrams: number;
  /**
   * @minimum 1
   * @nullable
   */
  lengthMm?: number | null;
  /**
   * @minimum 1
   * @nullable
   */
  widthMm?: number | null;
  /**
   * @minimum 1
   * @nullable
   */
  heightMm?: number | null;
  status: ProductVariantStatus;
  version: number;
  /** @nullable */
  effectivePrice?: string | null;
  /**
   * @nullable
   * @pattern ^[1-9][0-9]*$
   */
  effectivePriceId?: string | null;
  /** @nullable */
  effectivePriceVersion?: number | null;
  /** @nullable */
  bundle?: ProductVariantDtoBundle;
}